SKILL-219: Briefing-only triage before repair for lone unparseable_gate_failure
finalised subtask checkpoint on 'feat/SKILL-219-validate-gate-finding-structure'
phase=commit_push generation=0

Skill-Bill-Subtask: SKILL-219/2
