+++ .feature-specs/SKILL-220-principle-conformance-refactor/spec_subtask_5_oversized_goal_runner.md
@@ -60,7 +60,8 @@ makes unnecessary.
 5. No SQL or JDBC type escapes `runtime-infra-sqlite`, confirmed by existing
    type-leak architecture tests.
 6. Detekt complexity suppressions removed by the split are deleted rather
-   than moved; any retained suppression carries a reason.
+   than moved. Do not add new suppressions. Leftover `@Suppress` is
+   SKILL-221.
 7. `scripts/validate` passes.
 8. No test is added. If a split needs a new test to be safe, the split is
    wrong.
