+++ .feature-specs/SKILL-223-boundary-decision-write-hygiene/decomposition-manifest.yaml
@@ -1,0 +1,25 @@
+contract_version: "0.5"
+issue_key: "SKILL-223"
+feature_name: "boundary-decision-write-hygiene"
+parent_spec_path: ".feature-specs/SKILL-223-boundary-decision-write-hygiene/spec.md"
+execution_model: same_branch_commit_per_subtask
+base_branch: main
+feature_branch: feat/SKILL-223-boundary-decision-write-hygiene
+stack_branches: []
+current_subtask_intent:
+  subtask_id: 1
+  action: start
+subtasks:
+  - id: 1
+    name: "Decision write supersession hygiene"
+    spec_path: ".feature-specs/SKILL-223-boundary-decision-write-hygiene/spec_subtask_1_decision-write-supersession-hygiene.md"
+    status: pending
+    branch: null
+    commit_sha: null
+    workflow_id: null
+    blocked_reason: null
+    last_resumable_step: null
+    linear_issue_id: null
+    finalizing_agent_id: null
+    participating_agent_ids: []
+    dependencies: []
