+++ .feature-specs/done/SKILL-219-validate-gate-finding-structure/decomposition-manifest.yaml
@@ -3,14 +3,14 @@ contract_version: "0.5"
 issue_key: "SKILL-219"
 feature_name: "validate-gate-finding-structure"
 parent_spec_path: ".feature-specs/SKILL-219-validate-gate-finding-structure/spec.md"
-status: "in_progress"
+status: "complete"
 execution_model: "same_branch_commit_per_subtask"
 base_branch: "main"
 feature_branch: "feat/SKILL-219-validate-gate-finding-structure"
 stack_branches: []
 current_subtask_intent:
-  subtask_id: 2
-  action: "resume"
+  subtask_id: 0
+  action: "complete"
 subtasks:
 - id: 1
   name: "Discrete Gradle gate findings"
@@ -28,12 +28,12 @@ subtasks:
 - id: 2
   name: "Unparseable blob triage briefing"
   spec_path: ".feature-specs/SKILL-219-validate-gate-finding-structure/spec_subtask_2_unparseable-blob-triage-briefing.md"
-  status: "in_progress"
+  status: "complete"
   branch: null
   commit_sha: null
   workflow_id: "wftr-20260829-034137-hu1t"
   blocked_reason: null
-  last_resumable_step: null
+  last_resumable_step: "commit_push"
   linear_issue_id: null
   finalizing_agent_id: null
   participating_agent_ids: []
