+++ .feature-specs/done/SKILL-219-validate-gate-finding-structure/spec.md
