+++ .feature-specs/done/SKILL-219-validate-gate-finding-structure/spec_subtask_1_discrete-gradle-gate-findings.md
