+++ .feature-specs/done/SKILL-219-validate-gate-finding-structure/spec_subtask_2_unparseable-blob-triage-briefing.md
