+++ runtime-kotlin/runtime-application/src/main/kotlin/skillbill/application/featuretask/FeatureTaskRuntimePhasePromptValidateDirectives.kt
@@ -1,6 +1,9 @@
 package skillbill.application.featuretask
 
+import skillbill.application.featuretask.validation.model.ValidationFindingSetProjection
+import skillbill.application.model.FeatureTaskRuntimePhaseLaunchBriefing
 import skillbill.workflow.taskruntime.FeatureTaskRuntimePhaseWorkflowDefinition
+import skillbill.workflow.taskruntime.model.FeatureTaskRuntimeHandoffProjectionValue
 import skillbill.workflow.taskruntime.model.FeatureTaskRuntimePriorGapMemory
 
 private const val VALIDATE_PHASE_FORBIDDEN_EXTRAS: String =
@@ -130,4 +133,115 @@ internal fun phaseTaskDirective(phaseId: String, args: PhaseTaskDirectiveArgs = 
     return auditPhaseTaskDirective(args.priorGapMemory)
   }
   return phaseDirectives[phaseId] ?: error("No phase directive for runtime phase '$phaseId'.")
+}
+
+internal fun gateRepairNoOutputSchemaDirective(phaseId: String, triage: Boolean = false): String {
+  if (triage) {
+    return """
+      ## Gate triage — optional capture surface, no phase-output schema
+      This launch triages an unparseable gate blob before the first repair turn for the runtime-owned `$phaseId` gate.
+      Do not emit a Required final output JSON object, build_receipt, validation_receipt, gate_run_count, or any other
+      phase receipt or gate evidence. Do not spawn delegated subagents. Read the blob and cited paths.
+      When you can recommend a repair shape, you may emit produced_outputs.value (a JSON string) carrying
+      validation_repair_plan prose with suggested fields per item: item_id, module, rule_or_task, location,
+      failure_summary, fix_intent. Malformed or missing capture is fine; repair still runs without it.
+    """.trimIndent()
+  }
+  return """
+  ## Gate repair — prose only, no phase-output schema
+  This launch is a repair turn for the runtime-owned `$phaseId` gate. Do not emit a Required final
+  output JSON object, build_receipt, validation_receipt, gate_run_count, or any other phase envelope.
+  Do not spawn delegated subagents. Work in this single agent session in ordinary prose.
+
+  The runtime already ran the pack command and parsed the failures listed in this briefing. It will
+  re-run that command after you stop, and it may give you up to three repair turns against whatever
+  remains. Address every open finding in this turn — all at once, not one finding per turn.
+
+  Before editing, do brief reasoned planning in prose for each finding (or for a shared root cause
+  that covers several). Scale the plan to the finding:
+  - Small / obvious: a few lines of due diligence, then fix.
+  - Complex: a real short plan — blast radius, surrounding callers/contracts you checked, whether
+    the change can introduce new bugs, and how you will keep the fix local.
+
+  No defined plan schema. Do the thinking, then edit. When you are done fixing, stop.
+  Never silence findings with @Suppress, @file:Suppress, baselines, disabled rules, weakened
+  configuration, or skipped tests — fix the root cause instead.
+  """.trimIndent()
+}
+
+internal fun validationGateFindingsDirective(
+  phaseId: String,
+  findings: ValidationFindingSetProjection?,
+  triagePlan: String?,
+): String {
+  if (findings == null) return ""
+  val (sectionTitle, preamble) = when (phaseId) {
+    FeatureTaskRuntimePhaseWorkflowDefinition.PHASE_VALIDATE -> Pair(
+      "## Runtime validation gate findings",
+      "A prior gate run parsed these items. They are the full open set for this repair turn — fix " +
+        "every one in this session (shared root causes may collapse several into one change). Do not " +
+        "run `skill-bill validate`, `bill-code-check`, `./gradlew check`, `check " + "--" + "continue`, " +
+        "or the pack collect_all_full_gate_command. $VALIDATE_REPAIR_ALLOWED_TASKS" +
+        "Do not spawn delegated subagents.",
+    )
+    FeatureTaskRuntimePhaseWorkflowDefinition.PHASE_BUILD -> Pair(
+      "## Runtime build gate findings",
+      "A prior gate run parsed these items. They are the full open set for this repair turn — fix " +
+        "every one in this session (shared root causes may collapse several into one change). Run only " +
+        "the pack-declared build command when you need console detail. Do not run `skill-bill " +
+        "validate`, `bill-code-check`, `./gradlew check`, `check " + "--" + "continue`, or the pack " +
+        "collect_all_full_gate_command. Do not spawn delegated subagents.",
+    )
+    else -> return ""
+  }
+  val lines = buildList {
+    add(sectionTitle)
+    add(preamble)
+    findings.findings.forEachIndexed { index, finding ->
+      add(
+        "${index + 1}. module=${finding.module} id=${finding.ruleOrTestId} " +
+          "location=${finding.location ?: "<unknown>"} message=${finding.message}",
+      )
+    }
+    if (!triagePlan.isNullOrBlank()) {
+      add("## Triage working notes")
+      add(triagePlan)
+    }
+  }
+  return lines.joinToString("\n")
+}
+
+internal fun auditNoEarlierAuditLine(briefing: FeatureTaskRuntimePhaseLaunchBriefing): String =
+  if (briefing.priorGapMemory == null) {
+    "      Every audit re-checks every listed criterion from scratch against the tree, so there is no\n" +
+      "      earlier audit to account for and nothing to carry forward except the notes you emit now.\n"
+  } else {
+    "      Every audit re-checks every listed criterion from scratch against the tree; when this\n" +
+      "      briefing carries prior-gap memory, earlier audit value strings in prior_audit_values are\n" +
+      "      context you must account for, and a repeated criterion needs an explicit re-justification (below).\n"
+  }
+
+internal fun auditRoundScopeAddendum(briefing: FeatureTaskRuntimePhaseLaunchBriefing): String {
+  val memoryBlock = briefing.priorGapMemory?.let { memory ->
+    buildString {
+      append("\n      Prior-gap memory (round ${memory.round}): prior audit value strings:\n")
+      memory.priorAuditValues.forEach { value -> append("        - $value\n") }
+      append("      When a gap repeats a criterion already named in a prior audit value, require explicit\n")
+      append("      re-justification: name what the prior implement claimed and why the tree still fails it.\n")
+    }
+  }.orEmpty()
+  val auditProse = briefing.handoffEnvelope.projections
+    .firstOrNull { it.projectionName == "audit_prose" }
+    ?.fields
+    ?.firstOrNull { it.name == "value" }
+    ?.value
+    ?.let { (it as? FeatureTaskRuntimeHandoffProjectionValue.Text)?.text }
+  val scopeBlock = if (auditProse.isNullOrBlank()) {
+    ""
+  } else {
+    "\n      The previous audit value reported gaps in structured prose. Start there, then still decide " +
+      "every listed criterion from the tree: a repair can regress a criterion an earlier audit passed, " +
+      "and a narrow patch can open a new sibling gap."
+  }
+  return memoryBlock + scopeBlock
 }
