+++ runtime-kotlin/runtime-application/src/main/kotlin/skillbill/application/featuretask/validation/FeatureTaskRuntimeValidationGateCoordinator.kt
@@ -10,6 +10,7 @@ import skillbill.application.featuretask.validation.model.ValidationGateCycleReq
 import skillbill.application.featuretask.validation.model.ValidationGateCycleResult
 import skillbill.application.featuretask.validation.model.ValidationGateCycleTerminalOutcome
 import skillbill.application.featuretask.validation.model.ValidationGateProgressStore
+import skillbill.application.featuretask.validation.model.ValidationGateProgressWrite
 import skillbill.application.featuretask.validation.model.ValidationGateResolution
 import skillbill.application.featuretask.validation.model.ValidationGateTriageResult
 import skillbill.application.featuretask.validation.model.requiresUnparseableGateTriage
@@ -109,10 +110,13 @@ class FeatureTaskRuntimeValidationGateCoordinator(
     recordGateProgress(
       state = state,
       result = discovery,
-      findings = discoveryFindings,
-      repairWindowPhase = repairWindowPhaseFor(discoveryFindings),
-      repairsUsed = 0,
-      capturedTriagePlan = null,
+      write = ValidationGateProgressWrite(
+        repairWindowPhase = repairWindowPhaseFor(discoveryFindings),
+        remainingFindings = null,
+        completeFindings = discoveryFindings,
+        repairsUsed = 0,
+        capturedTriagePlan = null,
+      ),
     )
     if (discoveryFindings.isEmpty()) {
       return terminalCompletedResult(cycle.repositoryCheckpoint, measurements)
@@ -121,11 +125,13 @@ class FeatureTaskRuntimeValidationGateCoordinator(
     if (triagePlan != null) {
       persistProgress(
         state = state,
-        repairWindowPhase = FeatureTaskRuntimeValidationGateRepairWindowPhase.FINDINGS_OPEN,
-        remainingFindings = null,
-        completeFindings = discoveryFindings,
-        repairsUsed = 0,
-        capturedTriagePlan = triagePlan,
+        write = ValidationGateProgressWrite(
+          repairWindowPhase = FeatureTaskRuntimeValidationGateRepairWindowPhase.FINDINGS_OPEN,
+          remainingFindings = null,
+          completeFindings = discoveryFindings,
+          repairsUsed = 0,
+          capturedTriagePlan = triagePlan,
+        ),
       )
     }
     return repairLoop(
@@ -167,15 +173,16 @@ class FeatureTaskRuntimeValidationGateCoordinator(
       }
       val projection = ValidationFindingSetProjection(findings = currentFindings)
       if (repairsUsed >= MAX_REPAIR_TURNS) {
-      persistProgress(
-        state = state,
-        repairWindowPhase = FeatureTaskRuntimeValidationGateRepairWindowPhase.FINDINGS_OPEN,
-        remainingFindings = projection,
-        completeFindings = currentFindings,
-        repairsUsed = repairsUsed,
-        capturedTriagePlan = triagePlan,
-      )
-      return terminalBlockedResult(
+        persistProgress(
+          state = state,
+          write = ValidationGateProgressWrite.findingsOpen(
+            completeFindings = currentFindings,
+            repairsUsed = repairsUsed,
+            capturedTriagePlan = triagePlan,
+            remainingFindings = projection,
+          ),
+        )
+        return terminalBlockedResult(
           "Validation gate still reports ${currentFindings.size} finding(s) after $MAX_REPAIR_TURNS repair " +
             "turns; remaining findings are recorded for the operator.",
           remainingFindings = projection,
@@ -184,11 +191,11 @@ class FeatureTaskRuntimeValidationGateCoordinator(
       }
       persistProgress(
         state = state,
-        repairWindowPhase = FeatureTaskRuntimeValidationGateRepairWindowPhase.FINDINGS_OPEN,
-        remainingFindings = null,
-        completeFindings = currentFindings,
-        repairsUsed = repairsUsed,
-        capturedTriagePlan = triagePlan,
+        write = ValidationGateProgressWrite.findingsOpen(
+          completeFindings = currentFindings,
+          repairsUsed = repairsUsed,
+          capturedTriagePlan = triagePlan,
+        ),
       )
       when (val repair = cycle.agentRepairLauncher.launch(projection, repairsUsed + 1, triagePlan)) {
         is ValidationGateAgentRepairResult.Blocked -> return terminalBlockedResult(
@@ -198,21 +205,33 @@ class FeatureTaskRuntimeValidationGateCoordinator(
         )
         is ValidationGateAgentRepairResult.Completed -> repairsUsed++
       }
-      val verify = runGate(cycle, declaration, ValidationGateCyclePhase.POST_REPAIR_VERIFY)
-      val verifyFindings = findingsForRepairFromResult(verify)
-      recordGateProgress(
-        state = state,
-        result = verify,
-        findings = verifyFindings,
-        repairWindowPhase = repairWindowPhaseFor(verifyFindings),
-        repairsUsed = repairsUsed,
-        capturedTriagePlan = triagePlan,
-      )
-      if (verifyFindings.isEmpty()) {
+      currentFindings = verifyAfterRepair(state, declaration, repairsUsed, triagePlan)
+      if (currentFindings.isEmpty()) {
         return terminalCompletedResult(cycle.repositoryCheckpoint, measurements)
       }
-      currentFindings = verifyFindings
     }
+  }
+
+  private fun verifyAfterRepair(
+    state: ValidationGateCycleState,
+    declaration: ValidationGateDeclaration,
+    repairsUsed: Int,
+    triagePlan: String?,
+  ): List<ValidationGateFinding> {
+    val verify = runGate(state.cycle, declaration, ValidationGateCyclePhase.POST_REPAIR_VERIFY)
+    val verifyFindings = findingsForRepairFromResult(verify)
+    recordGateProgress(
+      state = state,
+      result = verify,
+      write = ValidationGateProgressWrite(
+        repairWindowPhase = repairWindowPhaseFor(verifyFindings),
+        remainingFindings = null,
+        completeFindings = verifyFindings,
+        repairsUsed = repairsUsed,
+        capturedTriagePlan = triagePlan,
+      ),
+    )
+    return verifyFindings
   }
 
   private fun repairWindowPhaseFor(
@@ -255,10 +274,7 @@ class FeatureTaskRuntimeValidationGateCoordinator(
   private fun recordGateProgress(
     state: ValidationGateCycleState,
     result: ValidationGateRunResult,
-    findings: List<ValidationGateFinding>,
-    repairWindowPhase: FeatureTaskRuntimeValidationGateRepairWindowPhase,
-    repairsUsed: Int,
-    capturedTriagePlan: String?,
+    write: ValidationGateProgressWrite,
   ) {
     state.measurements += FeatureTaskRuntimeValidationGateRunRecord(
       durationMs = result.durationMs,
@@ -266,32 +282,18 @@ class FeatureTaskRuntimeValidationGateCoordinator(
       cacheMode = result.cacheMode.wireValue,
       executedWorkUnits = result.executedWorkUnits,
     )
-    persistProgress(
-      state = state,
-      repairWindowPhase = repairWindowPhase,
-      remainingFindings = null,
-      completeFindings = findings,
-      repairsUsed = repairsUsed,
-      capturedTriagePlan = capturedTriagePlan,
-    )
+    persistProgress(state = state, write = write)
   }
 
-  private fun persistProgress(
-    state: ValidationGateCycleState,
-    repairWindowPhase: FeatureTaskRuntimeValidationGateRepairWindowPhase,
-    remainingFindings: ValidationFindingSetProjection?,
-    completeFindings: List<ValidationGateFinding>,
-    repairsUsed: Int,
-    capturedTriagePlan: String?,
-  ) {
+  private fun persistProgress(state: ValidationGateCycleState, write: ValidationGateProgressWrite) {
     val progress = FeatureTaskRuntimeValidationGateProgress(
       gateRunCount = state.measurements.size,
       gateRuns = state.measurements.toList(),
-      remainingFindings = remainingFindings?.toHandoffMaps().orEmpty(),
-      completeFindings = ValidationFindingSetProjection(findings = completeFindings).toHandoffMaps(),
-      repairWindowPhase = repairWindowPhase,
-      repairsUsed = repairsUsed,
-      capturedTriagePlan = capturedTriagePlan,
+      remainingFindings = write.remainingFindings?.toHandoffMaps().orEmpty(),
+      completeFindings = ValidationFindingSetProjection(findings = write.completeFindings).toHandoffMaps(),
+      repairWindowPhase = write.repairWindowPhase,
+      repairsUsed = write.repairsUsed,
+      capturedTriagePlan = write.capturedTriagePlan,
     )
     progressStore.persist(state.cycle.request.workflowId, progress, state.cycle.request.dbPathOverride)
     emitFeatureTaskRuntimeEventSafely(
