+++ runtime-kotlin/runtime-application/src/test/kotlin/skillbill/application/featuretask/validation/FeatureTaskRuntimeBuildGateFullDiscoverTest.kt
@@ -45,7 +45,7 @@ class FeatureTaskRuntimeBuildGateFullDiscoverTest {
         repositoryCheckpoint = "checkpoint",
         agentRepairLauncher = ValidationGateAgentRepairLauncher { findings, _, _ ->
           repairRuleIds += findings.findings.map { it.ruleOrTestId }
-          completedRepair(findings.findings)
+          completedRepair()
         },
       ),
     )
