+++ runtime-kotlin/runtime-infra-fs/src/main/kotlin/skillbill/infrastructure/fs/validation/FileSystemValidationGateGradlePathSupport.kt
@@ -1,0 +1,56 @@
+package skillbill.infrastructure.fs.validation
+
+import java.nio.file.Files
+import java.nio.file.Path
+
+internal object FileSystemValidationGateGradlePathSupport {
+  fun repoRelativeQualityPath(repo: Path, rawPath: String): String {
+    val diagnosticPath = Path.of(rawPath.removePrefix("file://"))
+    val canonicalRepo = canonicalizeExisting(repo)
+    return repoRelativeCompilerPath(repo, canonicalRepo, diagnosticPath, rawPath)
+  }
+
+  fun repoRelativeCompilerPath(repo: Path, canonicalRepo: Path, diagnosticPath: Path, rawPath: String): String {
+    val absolute = diagnosticPath.toAbsolutePath().normalize()
+    val canonicalFile = canonicalizeMaybeMissing(absolute)
+    if (canonicalFile.startsWith(canonicalRepo)) {
+      return canonicalRepo.relativize(canonicalFile).toString().replace('\\', '/')
+    }
+    if (absolute.startsWith(repo)) {
+      return repo.relativize(absolute).toString().replace('\\', '/')
+    }
+    return rawPath.replace('\\', '/').removePrefix("/")
+  }
+
+  fun canonicalizeExisting(path: Path): Path = runCatching { path.toRealPath() }.getOrDefault(path)
+
+  fun parseGradleTaskPath(taskPath: String): Pair<String, String> {
+    val segments = taskPath.trim(':').split(':').filter { it.isNotEmpty() }
+    return when (segments.size) {
+      0 -> "" to ""
+      1 -> "" to segments[0]
+      else -> segments.dropLast(1).joinToString(":") to segments.last()
+    }
+  }
+
+  fun filePathFromAdviceBlock(line: String): String? {
+    val match = Regex("""(?:file://)?(/[^\s:]+\.(?:kt|kts|java|gradle))""").find(line) ?: return null
+    return match.groupValues[1]
+  }
+
+  fun repoRelativeAdvicePath(rawPath: String): String = rawPath.removePrefix("file://").trimStart('/')
+
+  private fun canonicalizeMaybeMissing(path: Path): Path {
+    if (Files.exists(path)) {
+      return canonicalizeExisting(path)
+    }
+    val tail = ArrayDeque<Path>()
+    var current: Path? = path
+    while (current != null && !Files.exists(current)) {
+      current.fileName?.let(tail::addFirst)
+      current = current.parent
+    }
+    val realBase = current?.let(::canonicalizeExisting) ?: return path
+    return tail.fold(realBase) { acc, name -> acc.resolve(name) }
+  }
+}
